feat: implement movie detail fragment

+++ app/src/main/java/ru/ksart/tmdb_movies/data/network/TmdbApiService.kt
@@ -19,7 +19,7 @@ interface TmdbApiService {
 
     @GET("movie/{movie_id}")
     suspend fun getMovieById(
-        @Path("movie_id") movieId: Int,
+        @Path("movie_id") movieId: Long,
         @Query(PARAMS_LANGUAGE) language: String,
     ): Movie
