+++ app/src/main/java/ru/ksart/tmdb_movies/domain/repository/MoviesRepository.kt
@@ -7,7 +7,7 @@ import ru.ksart.tmdb_movies.data.entitis.MovieTop
 
 interface MoviesRepository {
 
-    suspend fun getMovieById(id: Int, language: String): Movie
+    suspend fun getMovieById(id: Long, language: String): Movie
 
     fun getMoviesTopRated(language: String): Flow<PagingData<MovieTop>>
 }
